fixed some inconsistencies in the classes, added grade class

+++ MathWiz/MathWiz/PracticeTest.cs
@@ -8,17 +8,19 @@ namespace MathWiz
 {
     public class PracticeTest : Test
     {
-        public short MinLevel { get; set; }
-        public short MaxLevel { get; set; }
+        public int MinLevel { get; set; }
+        public int MaxLevel { get; set; }
+        
 
         //empty constructor
         public PracticeTest() : base() { }
 
         //full constructor
-        public PracticeTest(List<Question> questions, TimeSpan timeLimit, short minLevel, short maxLevel) : base(questions, timeLimit)
+        public PracticeTest(List<Question> questions, TimeSpan timeLimit, int minLevel, int maxLevel) : base(questions, timeLimit)
         {
             this.MinLevel = minLevel;
             this.MaxLevel = maxLevel;
+            
         }
     }
 }
